Add files via upload
The dataset consists of 34 feature attributes and one target attribute, making a total of 351 (train + test) instances. The feature attributes are primarily numerical, representing various signal properties captured by the radar. The target attribute is binary, indicating whether the radar return shows evidence of some form of structure in the ionosphere (1 for good) or not (0 for bad). The primary task here is binary classification.

Used Random Forest Classifier to predict. For accuracy testing used Confusion Matrix and Classification report.